Add gamemode
Finally putting this up, needed to have done this a long time ago

Fully fledged garry's mod gamemode that leans on ACF-3
Similar to Squad's AAS/RAAS gamemode, each team must capture flags to start reducing the enemy's tickets to 0. Depending on flag status when captured, the capturing team will game tickets, and possibly reduce the enemy's tickets by an amount (if it was previously owned by them)

Each player has a resource known as Requisition, which is passively gained over time at a rate determined by map settings as well as how they have behaved throughout the match, as in karma. If they have been teamkilling or lingering around the enemy safezone, they will lose karma considerably. The only way to gain karma back is to do something productive for the team as a whole, which is to capture points.

Map setup is a breeze with a built-in STool which is only available to admins when the server is in editmode